+++ Decsys/ClientApp/src/animations/fade-out.js
@@ -1,0 +1,13 @@
+import { keyframes } from "styled-components";
+
+export const fadeOut = keyframes`
+  from {
+    opacity: 1;
+  }
+
+  to {
+    opacity: 0;
+  }
+`;
+
+export default fadeOut;
