+++ Decsys/ClientApp/src/animations/rotate.js
@@ -1,15 +1,5 @@
 import { keyframes } from "styled-components";
 
-export const fadeOut = keyframes`
-  from {
-    opacity: 1;
-  }
-
-  to {
-    opacity: 0;
-  }
-`;
-
 export const rotate = keyframes`
   from {
     transform: rotate(0deg);
@@ -19,3 +9,5 @@ export const rotate = keyframes`
     transform: rotate(360deg);
   }
 `;
+
+export default rotate;
