+++ Decsys/ClientApp/src/components/EditorBar/index.js
@@ -1,3 +1,4 @@
 import EditorBar from "./EditorBar";
 
+export { default as EditorBarContext } from "./Context";
 export default EditorBar;
