--- Decsys/ClientApp/src/components/ExportModal/index.js
@@ -1,2 +1,0 @@
-import ExportModal from "./ExportModal";
-export default ExportModal;
